API Reference

API Reference

ExpressBase module

ExpressBase.Files module

ExpressBase module

ExpressBase.Calculation — Type

Represent all materials calculations.

ExpressBase.Action — Type

Represent an atomic action for a specific Calculation type.

ExpressBase.calculation — Function

calculation(::Action)

Return the calculation type of the Action.

ExpressBase.Files module

ExpressBase.Files.save — Function

save(file, data)

Save data to file.

By now, YAML, JSON, and TOML formats are supported. The format is recognized by the file extension.

If data is a Dict, its keys should be Strings so that load can return the same data.

Warning

Allowed data types can be referenced in JSON.jl documentation and YAML.jl documentation. For TOML format, only AbstractDict type is allowed.

ExpressBase.Files.load — Function

load(file)

Load data from file to a Dict.

By now, YAML, JSON, and TOML formats are supported. The format is recognized by the file extension.

ExpressBase.Files.of_format — Function

of_format(destination, source)

Convert source to the format of destination. Similar to oftype.